requirements.txt
pandas
chromadb
openai
umap-learn
plotly
scikit-learn